Changement de noms : Avant           Maintenant Group NDF     = NDF NDF           = NDFL
Tests à compléter.
