--- modules/group-ndf/group-ndf.config.json
@@ -1,13 +1,0 @@
-{
-	"name": "Group NDF",
-	"slug": "group-ndf",
-	"version": "1.0.0.0",
-	"description": "Le module gérant les Groupe NDF",
-	"path": "modules/group-ndf/",
-	"dependencies": {
-		"model": {},
-		"class": {},
-		"action": {},
-		"helper": {}
-	}
-}
