+++ modules/ndf/action/ndf.action.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Classe gérant les actions NDF
+ * Classe gérant les actions des notes de frais.
  *
  * @author eoxia
  * @since 1.0.0.0
@@ -12,27 +12,43 @@
 
 namespace note_de_frais;
 
-if ( ! defined( 'ABSPATH' ) ) { exit; }
+if ( ! defined( 'ABSPATH' ) ) {
+	exit;
+}
 
 /**
- * Classe gérant les actions NDF
+ * Classe gérant les notes de frais.
  */
 class NDF_Action {
 
+	/**
+	 * Le constructeur
+	 *
+	 * @since 1.0.0.0
+	 * @version 1.0.0.0
+	 */
 	public function __construct() {
-		add_action( 'wp_ajax_open_note_de_frais', array( $this, 'callback_open_note_de_frais' ) );
-		add_action( 'wp_ajax_modify_note_de_frais', array( $this, 'callback_modify_note_de_frais' ) );
-		add_action( 'wp_ajax_delete_note_de_frais', array( $this, 'callback_delete_note_de_frais' ) );
-		add_action( 'wp_ajax_add_note_de_frais', array( $this, 'callback_add_note_de_frais' ) );
+		add_action( 'wp_ajax_open_ndf', array( $this, 'callback_open_ndf' ) );
+		add_action( 'wp_ajax_create_ndf', array( $this, 'callback_create_ndf' ) );
+		add_action( 'wp_ajax_modify_ndf', array( $this, 'callback_modify_ndf' ) );
+		add_action( 'wp_ajax_archive_ndf', array( $this, 'callback_archive_ndf' ) );
+		add_action( 'wp_ajax_export_ndf', array( $this, 'callback_export_ndf' ) );
 	}
 
-	public function callback_open_note_de_frais() {
-		check_ajax_referer( 'open_note_de_frais' );
+	/**
+	 * Action : Ouvrir une note de frais.
+	 *
+	 * @param  string $_wpnonce          Nonce 'open_ndf' for check.
+	 * @return string $view              Json NDF_Class::display(ID).
+	 * @return string $callback_success  Json callback noteDeFrais.NDF.openNdf().
+	 */
+	public function callback_open_ndf() {
+		check_ajax_referer( 'open_ndf' );
 
-		$group_id = isset( $_POST['group_id'] ) ? intval( $_POST['group_id'] ) : -1;
+		$ndf_id = isset( $_POST['id'] ) ? intval( $_POST['id'] ) : -1;
 
 		ob_start();
-		NDF_Class::g()->display( $group_id );
+		NDF_Class::g()->display( $ndf_id );
 		$response = ob_get_clean();
 
 		wp_send_json_success( array(
@@ -43,72 +59,200 @@ class NDF_Action {
 		) );
 	}
 
-	public function callback_modify_note_de_frais() {
-		check_ajax_referer( 'modify_note_de_frais' );
+	/**
+	 * Action : créer une note de frais.
+	 *
+	 * @param  string $_wpnonce          Nonce 'create_ndf' for check.
+	 * @return string $view              Json NDF_Class::display(ID).
+	 * @return string $main_view         Json NDF_Class::display().
+	 * @return string $callback_success  Json callback noteDeFrais.NDF.openNdf().
+	 */
+	public function callback_create_ndf() {
+		check_ajax_referer( 'create_ndf' );
 
-		$group_id = isset( $_POST['group_id'] ) ? intval( $_POST['group_id'] ) : -1;
+		$user = User_Class::g()->get( array(
+			'include' => get_current_user_id(),
+		), true );
 
-		if ( isset( $_POST['row'] ) ) {
-			foreach ( $_POST['row'] as $row ) {
-				$row['post_parent'] = $group_id;
-				$current_row = NDF_Class::g()->update( $row );
-				$all_new_row[] = $current_row->id;
+		$date = current_time( 'Y-m' );
+
+		$identifier = get_user_meta( get_current_user_id(), 'ndf_' . $date . '_identifier', true );
+
+		if ( empty( $identifier ) ) {
+			$identifier = 001;
+		} else {
+			$identifier++;
+		}
+
+		if ( intval( strlen( $identifier ) ) === 1 ) {
+			$identifier = '00' . $identifier;
+		}
+
+		if ( intval( strlen( $identifier ) ) === 2 ) {
+			$identifier = '0' . $identifier;
+		}
+
+		$ndf = NDF_Class::g()->update( array(
+			'post_title' => strtoupper( $user->login ) . '-' . $date . '-' . $identifier,
+			'post_status' => 'publish',
+		) );
+
+		update_user_meta( get_current_user_id(), 'ndf_' . $date . '_identifier', $identifier );
+
+		ob_start();
+		NDF_Class::g()->display( $ndf->id );
+		$response = ob_get_clean();
+
+		ob_start();
+		NDF_Class::g()->display();
+		$response_main_view = ob_get_clean();
+
+		wp_send_json_success( array(
+			'namespace'        => 'noteDeFrais',
+			'module'           => 'NDF',
+			'callback_success' => 'openNdf',
+			'view'             => $response,
+			'main_view'        => $response_main_view,
+		) );
+	}
+
+	/**
+	 * Action : modifier une note de frais.
+	 *
+	 * @param  string $_wpnonce          Nonce 'modify_ndf' for check.
+	 * @param  string Mixed              @see NDF_Model.
+	 * @return string $ndf               Json updated ndf.
+	 * @return string $view              Json NDF_Class::display(ID).
+	 * @return string $callback_success  Json callback noteDeFrais.NDF.refreshNDF().
+	 */
+	public function callback_modify_ndf() {
+		check_ajax_referer( 'modify_ndf' );
+
+		$ndf = NDF_Class::g()->update( $_POST );
+
+		ob_start();
+		NDF_Class::g()->display( $ndf->id );
+		$response = ob_get_clean();
+
+		wp_send_json_success( array(
+			'namespace' => 'noteDeFrais',
+			'module' => 'NDF',
+			'callback_success' => 'refreshNDF',
+			'ndf' => $ndf,
+			'view' => $response,
+		) );
+	}
+
+	/**
+	 * Action : archiver une note de frais.
+	 *
+	 * @param  string $_wpnonce          Nonce 'archive_ndf' for check.
+	 * @param  string $id                ID de le note de frais.
+	 * @return string $ndf             Json updated note de frais.
+	 * @return string $callback_success  Json callback noteDeFrais.NDF.archived().
+	 */
+	public function callback_archive_ndf() {
+		check_ajax_referer( 'archive_ndf' );
+
+		$id = ! empty( $_POST['id'] ) ? (int) $_POST['id'] : 0;
+
+		if ( empty( $id ) ) {
+			wp_send_json_error();
+		}
+
+		$ndf = NDF_Class::g()->get( array(
+			'post__in' => array( $id ),
+		), true );
+
+		$ndf->status = 'archive';
+
+		wp_send_json_success( array(
+			'namespace' => 'noteDeFrais',
+			'module' => 'NDF',
+			'callback_success' => 'archived',
+			'ndf' => NDF_Class::g()->update( $ndf ),
+		) );
+	}
+
+	/**
+	 * Génère un document .odt avec les données qui vont bien.
+	 *
+	 * @param  string $_wpnonce          Nonce 'export_ndf' for check.
+	 * @param  string $id                ID de le note de frais.
+	 * @return string $link              Json link file.
+	 * @return string $filename          Json filename.
+	 * @return string $callback_success  Json callback noteDeFrais.NDF.exportedNoteDeFraisSuccess().
+	 *
+	 * @since 1.0.0.0
+	 * @version 1.0.0.0
+	 */
+	public function callback_export_ndf() {
+		check_ajax_referer( 'export_ndf' );
+
+		$total_tax_inclusive_amount = 0;
+		$total_tax_amount = 0;
+		$ndf = NDF_Class::g()->get( array(
+			'include' => array( $_POST['id'] ),
+		), true );
+
+		$ndfls = NDFL_Class::g()->get( array(
+			'post_parent' => $_POST['id'],
+		) );
+
+		$user = User_Class::g()->get( array(
+			'include' => array( $ndf->author_id ),
+		), true );
+
+		$sheet_details = array(
+			'ndf' => array(
+				'type' => 'segment',
+				'value' => array(),
+			),
+		);
+
+		$periode = explode( '-', $ndf->title );
+		$periode = $periode[2] . '/' . $periode[1];
+
+		if ( ! empty( $user->firstname ) && ! empty( $user->lastname ) ) {
+			$sheet_details['utilisateur_prenom_nom'] = $user->firstname . ' ' . $user->lastname;
+		}
+		$sheet_details['utilisateur_email'] = $user->email;
+		$sheet_details['periode'] = $periode;
+
+		if ( empty( $sheet_details['utilisateur_prenom_nom'] ) ) {
+			$sheet_details['utilisateur_prenom_nom'] = $user->login;
+		}
+
+		$sheet_details['status'] = $ndf->validation_status;
+		$sheet_details['miseajour'] = $ndf->date_modified;
+
+		if ( ! empty( $ndfls ) ) {
+			foreach ( $ndfls as $ndfl ) {
+				$sheet_details['ndf']['value'][] = array(
+					'date' => $ndfl->date,
+					'libelle' => $ndfl->title,
+					'km' => $ndfl->distance,
+					'ttc' => $ndfl->tax_inclusive_amount . '€',
+					'tva' => $ndfl->tax_amount . '€',
+					'tvarecup' => 'Je n\'existe pas',
+					'photo' => 'photo',
+				);
+
+				$total_tax_inclusive_amount += $ndfl->tax_inclusive_amount;
+				$total_tax_amount += $ndfl->tax_amount;
 			}
 		}
 
-		ob_start();
-		NDF_Class::g()->display( $group_id );
-		$response = ob_get_clean();
+		$sheet_details['totaltva'] = $total_tax_amount . '€';
+		$sheet_details['totalttc'] = $total_tax_inclusive_amount . '€';
 
-		$group = Group_NDF_Class::g()->get( array(
-			'post__in' => array( $group_id ),
-		), true );
-
+		$response = document_class::g()->create_document( $ndf, $sheet_details );
 		wp_send_json_success( array(
 			'namespace' => 'noteDeFrais',
 			'module' => 'NDF',
-			'callback_success' => 'refreshNDF',
-			'group' => $group,
-			'view' => $response,
-		) );
-	}
-
-	public function callback_delete_note_de_frais() {
-		check_ajax_referer( 'delete_note_de_frais' );
-
-		$group_id = isset( $_POST['group_id'] ) ? intval( $_POST['group_id'] ) : -1;
-		$row_to_delete = isset( $_POST['ndf_id'] ) ? intval( $_POST['ndf_id'] ) : -1;
-
-		$row = NDF_Class::g()->get( array(
-			'post__in' => array( $row_to_delete ),
-		), true );
-
-		$row->status = 'trash';
-		NDF_Class::g()->update( $row );
-
-		ob_start();
-		NDF_Class::g()->display( $group_id );
-		$response = ob_get_clean();
-
-		wp_send_json_success( array(
-			'namespace' => 'noteDeFrais',
-			'module' => 'NDF',
-			'callback_success' => 'refreshNDF',
-			'view' => $response,
-		) );
-	}
-
-	public function callback_add_note_de_frais() {
-		check_ajax_referer( 'add_note_de_frais' );
-
-		$group_id = isset( $_POST['group_id'] ) ? intval( $_POST['group_id'] ) : -1;
-		NDF_Class::g()->create( $_POST );
-
-		wp_send_json_success( array(
-			'namespace' => '',
-			'module' => '',
-			'callback_success' => '',
-			'response' => NDF_Class::g()->display( $group_id ),
+			'link' => $response['link'],
+			'filename' => $response['filename'],
+			'callback_success' => 'exportedNoteDeFraisSuccess',
 		) );
 	}
 }
