+++ modules/ndf/class/ndf.class.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Classe gérant les NDF
+ * Classe gérant les notes de frais.
  *
  * @author eoxia
  * @since 1.0.0.0
@@ -12,10 +12,12 @@
 
 namespace note_de_frais;
 
-if ( ! defined( 'ABSPATH' ) ) { exit; }
+if ( ! defined( 'ABSPATH' ) ) {
+	exit;
+}
 
 /**
- * Classe gérant les groupe NDF
+ * Classe gérant les notes de frais.
  */
 class NDF_Class extends \eoxia\Post_Class {
 
@@ -24,64 +26,43 @@ class NDF_Class extends \eoxia\Post_Class {
 	 *
 	 * @var string
 	 */
-	protected $model_name                   = '\note_de_frais\NDF_Model';
+	protected $model_name = '\note_de_frais\NDF_Model';
 
 	/**
 	 * Le post type
 	 *
 	 * @var string
 	 */
-	protected $post_type                    = 'ndf';
+	protected $post_type  = 'ndf';
 
 	/**
 	 * La clé principale du modèle
 	 *
 	 * @var string
 	 */
-	protected $meta_key                     = '_ndf';
-
-	/**
-	 * Fonction de callback avant d'insérer les données en mode POST.
-	 *
-	 * @var array
-	 */
-	protected $before_post_function = array( '\eoxia\convert_date_time', '\note_de_frais\before_update_ndf' );
-
-
-	/**
-	 * Fonction de callback avant de mêttre à jour les données en mode PUT.
-	 *
-	 * @var array
-	 */
-	protected $before_put_function = array( '\eoxia\convert_date_time', '\note_de_frais\before_update_ndf' );
+	protected $meta_key   = '_ndf';
 
 	/**
 	 * La fonction appelée automatiquement avant la création de l'objet dans la base de donnée
 	 *
 	 * @var array
 	 */
-	protected $after_post_function = array( '\note_de_frais\after_update_ndf' );
+	protected $before_post_function = array( '\eoxia\convert_date_time' );
 
 	/**
 	 * La fonction appelée automatiquement avant la modification de l'objet dans la base de donnée
 	 *
 	 * @var array
 	 */
-	protected $after_put_function = array( '\note_de_frais\after_update_ndf' );
-
-	/**
-	 * La fonction appelée automatiquement avant la modification de l'objet dans la base de donnée
-	 *
-	 * @var array
-	 */
-	protected $after_get_function = array( '\eoxia\construct_current_date_time' );
+	protected $before_put_function = array( '\eoxia\convert_date_time' );
 
 	/**
 	 * La fonction appelée automatiquement après la récupération de l'objet dans la base de donnée
 	 *
 	 * @var array
 	 */
-	// protected $after_get_function = array( '\note_de_frais\sanitize_ndf_class' );
+	protected $after_get_function = array( '\eoxia\construct_current_date_time', '\note_de_frais\get_full_ndf' );
+
 	/**
 	 * Le nom pour le resgister post type
 	 *
@@ -89,17 +70,48 @@ class NDF_Class extends \eoxia\Post_Class {
 	 */
 	protected $post_type_name = 'NDF';
 
-	public function display( $group_id = -1 ) {
-		$notes = $this->get( array(
-			'post_parent' => $group_id,
+	/**
+	 * Nom du statut à afficher.
+	 *
+	 * @var array
+	 */
+	public $status = array(
+		'En cours' => 'en-cours',
+		'Validée'  => 'valide',
+		'Payée'    => 'paye',
+		'Refusée'  => 'refuse',
+	);
+
+	/**
+	 * Récupères les notes de frais et les envoies à la vue principale.
+	 *
+	 * @param  array  $status Post_status, permet d'afficher notes archivés ou publique.
+	 * @return void
+	 *
+	 * @since 1.0.0.0
+	 * @version 1.0.0.0
+	 */
+	public function display( $status = array( 'publish', 'future' ) ) {
+		$ndfs = $this->get( array(
+			'post_status' => $status,
 		) );
-		$template_vars = array(
-			'notes' => $notes,
-		);
-		$template_vars['group'] = Group_NDF_Class::g()->get( array(
-			'id' => $group_id,
-		), true );
-		\eoxia\View_Util::exec( 'note-de-frais', 'ndf', 'main', $template_vars );
+		\eoxia\View_Util::exec( 'note-de-frais', 'ndf', 'main', array(
+			'ndfs' => $ndfs,
+			'status' => $status,
+		) );
+	}
+
+	/**
+	 * Récupère le nom de statut avec en fonction du code.
+	 *
+	 * @param  string $status Code statut @see this->status.
+	 * @return string         Nom du statut @see this->status.
+	 *
+	 * @since 1.0.0.0
+	 * @version 1.0.0.0
+	 */
+	public function get_status( $status ) {
+		return $this->status[ $status ];
 	}
 }
