+++ modules/ndf/ndf.config.json
@@ -2,7 +2,7 @@
 	"name": "NDF",
 	"slug": "ndf",
 	"version": "1.0.0.0",
-	"description": "Le module gérant les NDF",
+	"description": "Le module gérant les notes de frais",
 	"path": "modules/ndf/",
 	"dependencies": {
 		"model": {},
